Testing out PyomoModelHandler
